CISMA_public/requirements.txt
compressai==1.2.0
einops==0.8.0
imageio==2.30.0
matplotlib==3.7.1
numpy==1.23.4
omegaconf==2.3.0
opencv_python==4.8.0.74
opencv_python_headless==4.8.1.78
Pillow==9.5.0
Pillow==11.1.0
pytorch_msssim==1.0.0
PyYAML==6.0
PyYAML==6.0.2
scikit_learn==1.2.2
thop==0.1.1.post2209072238
timm==0.9.7
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.65.0
